pyproject.toml
[tool.poetry]
name = "cloud-drift-analyzer"
version = "0.1.0"
description = "A tool for analyzing infrastructure drift between IaC and cloud providers"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cloud_drift_analyzer"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.6"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
boto3 = "^1.34"
google-cloud-resource-manager = "^1.12"
google-cloud-storage = "^2.14"
pulumi = "^3.0"
python-terraform = "^0.10"
slack-sdk = "^3.27"
jinja2 = "^3.1"
python-crontab = "^3.0"
aiohttp = "^3.9"
sqlalchemy = "^2.0"
sqlmodel = "^0.0.14"
alembic = "^1.13"
python-jose = {extras = ["cryptography"], version = "^3.3"}
python-multipart = "^0.0.9"
rich = "^13.7"
typer = "^0.9"
structlog = "^24.1"
deepdiff = "^6.7.1"
aiosqlite = "^0.19.0"
aioboto3 = "^12.3.0"  # Added aioboto3
PyJWT = "^2.8.0"  # Added PyJWT for OIDC token validation
passlib = {extras = ["bcrypt"], version = "^1.7.4"}  # Added for password hashing
greenlet = "^3.2.2"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
black = "^24.1"
isort = "^5.13"
mypy = "^1.8"
flake8 = "^7.0"
moto = "^4.2"
faker = "^22.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cloud-drift = "cloud_drift_analyzer.cli.main:app"
